Observer tervezési minta javítása, implementálása

+++ src/main/java/com/example/sorozatok/service/MovieManager.java
@@ -1,12 +1,15 @@
 package com.example.sorozatok.service;
 import com.example.sorozatok.model.Film;
+import com.example.sorozatok.observer.Observable;
+import com.example.sorozatok.observer.Observer;
 import com.example.sorozatok.repository.FilmRepository;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
-public class MovieManager {
+public class MovieManager extends Observable<Observer> {
     private final FilmRepository filmRepository;
 
     public MovieManager(FilmRepository filmRepository) {
@@ -24,10 +27,16 @@ public class MovieManager {
 
     public void deleteFilm(int filmId) {
         try {
-            filmRepository.delete(filmId);
+            Optional<Film> filmToDelete = filmRepository.findById(filmId);
+            if (filmToDelete.isPresent()) {
+                filmRepository.delete(filmId);
+                notifyObservers(observer -> observer.onMovieDeleted(filmToDelete));
+            }
         } catch (SQLException e) {
             e.printStackTrace();
         }
+
+
     }
 
     public List<Film> filterFilms(List<Film> films, String filterText) {
